requirements.txt
altair==5.4.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.2
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.2
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
streamlit==1.38.0
streamlit-authenticator==0.3.3
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
